refactor(parser): Refactor node flags and checkpoint benchmarks
- Replace direct node flag access with helpers across parser, GLR, incremental, and rewrite paths so named, extra, missing, and error state stays consistent.
- Add a no-tree benchmark mode that preserves external-scanner checkpoints so checkpoint overhead can be measured separately from tree materialization.
- Tighten regression coverage for arena reset, dirty-state clearing, and checkpoint retention to catch stale state leaks.

+++ parser_result_node_helpers.go
@@ -64,7 +64,7 @@ func replaceChildRangeWithSingleNode(parent *Node, start, end int, replacement *
 			continue
 		}
 		child.parent = parent
-		child.childIndex = i
+		child.childIndex = int32(i)
 	}
 }
